+++ app/src/main/java/com/ivy/wallet/domain/action/exchange/ExchangeAct.kt
@@ -6,13 +6,13 @@ import com.ivy.frp.then
 import com.ivy.wallet.domain.pure.exchange.ExchangeData
 import com.ivy.wallet.domain.pure.exchange.exchange
 import com.ivy.wallet.io.persistence.dao.ExchangeRateDao
-import java.math.BigDecimal
+
 import javax.inject.Inject
 
 class ExchangeAct @Inject constructor(
     private val exchangeRateDao: ExchangeRateDao,
-) : FPAction<ExchangeAct.Input, Option<BigDecimal>>() {
-    override suspend fun Input.compose(): suspend () -> Option<BigDecimal> = suspend {
+) : FPAction<ExchangeAct.Input, Option<Double>>() {
+    override suspend fun Input.compose(): suspend () -> Option<Double> = suspend {
         exchange(
             data = data,
             amount = amount,
@@ -24,13 +24,13 @@ class ExchangeAct @Inject constructor(
 
     data class Input(
         val data: ExchangeData,
-        val amount: BigDecimal
+        val amount: Double
     )
 }
 
 fun actInput(
     data: ExchangeData,
-    amount: BigDecimal
+    amount: Double
 ): ExchangeAct.Input = ExchangeAct.Input(
     data = data,
     amount = amount
