+++ app/src/main/java/com/ivy/wallet/domain/action/settings/CalcBufferDiffAct.kt
@@ -1,17 +1,17 @@
 package com.ivy.wallet.domain.action.settings
 
 import com.ivy.frp.action.FPAction
-import java.math.BigDecimal
+
 import javax.inject.Inject
 
-class CalcBufferDiffAct @Inject constructor() : FPAction<CalcBufferDiffAct.Input, BigDecimal>() {
+class CalcBufferDiffAct @Inject constructor() : FPAction<CalcBufferDiffAct.Input, Double>() {
 
-    override suspend fun Input.compose(): suspend () -> BigDecimal = {
+    override suspend fun Input.compose(): suspend () -> Double = {
         balance - buffer
     }
 
     data class Input(
-        val balance: BigDecimal,
-        val buffer: BigDecimal
+        val balance: Double,
+        val buffer: Double
     )
 }
