+++ app/src/main/java/com/ivy/wallet/domain/data/core/Settings.kt
@@ -3,13 +3,13 @@ package com.ivy.wallet.domain.data.core
 import com.ivy.design.l0_system.Theme
 import com.ivy.wallet.io.network.data.SettingsDTO
 import com.ivy.wallet.io.persistence.data.SettingsEntity
-import java.math.BigDecimal
+
 import java.util.*
 
 data class Settings(
     val theme: Theme,
     val baseCurrency: String,
-    val bufferAmount: BigDecimal,
+    val bufferAmount: Double,
     val name: String,
 
     val id: UUID = UUID.randomUUID()
